+++ pyramid_persona/templates/persona.js
@@ -17,7 +17,11 @@ $(function() {
 		    csrf_token: '%(csrf_token)s'
 		},
 		dataType: 'json',
-		success: function(res, status, xhr) { window.location = res['redirect']; },
+		success: function(res, status, xhr) {
+		    if(!res['success'])
+			navigator.id.logout();
+		    window.location = res['redirect'];
+		},
 		error: function(xhr, status, err) {
 		    navigator.id.logout();
 		    alert("Login failure: " + err);