+++ pyramid_persona/views.py
@@ -27,7 +27,7 @@ def login(request):
     """View to check the persona assertion and remember the user"""
     email = verify_login(request)
     request.response.headers.extend(remember(request, email))
-    return {'redirect': request.POST['came_from']}
+    return {'redirect': request.POST['came_from'], 'success': True}
 
 
 def logout(request):